execute code sometimes when drag-and-drop related events happen
The thing previously missing was to attach the listener to every subwindow of
the scroll pane in question as well as the scroll pane itself. Still unknown:
how to make this happen in production not a script. Do we have to attach
listeners to each document? What happens if a new document is created? etc.

This appears only to work to accept a link dragged in after it has first
rejected an invalid drop of some plain text.